+++ test/meshes_interface.jl
@@ -1,7 +1,7 @@
 @testsnippet InterfacesSetup begin
     using Meshes
     using CountriesBorders
-    using CountriesBorders.GeoBasics: to_multi, bboxes, polyareas, to_cart_point
+    using CountriesBorders.GeoBasics: to_multi, bboxes, polyareas, to_cartesian_point
     using CountriesBorders.BasicTypes: valuetype
     using CoordRefSystems
     using Test
@@ -30,11 +30,11 @@ end
 
 @testitem "in_exit_early" setup=[InterfacesSetup] begin
     using CountriesBorders: DOMAIN
-    using CountriesBorders.GeoBasics: in_exit_early, to_multi, to_cart_point
+    using CountriesBorders.GeoBasics: in_exit_early, to_multi, to_cartesian_point
     using CoordRefSystems: CoordRefSystems, Cartesian
     dmn = extract_countries("*")
     # Add consistency checks with the standard `in` method not doing exit early
-    _in(p, cb::CountryBorder) = in(to_cart_point(valuetype(cb), p), to_multi(Cartesian, cb))
+    _in(p, cb::CountryBorder) = in(to_cartesian_point(valuetype(cb), p), to_multi(Cartesian, cb))
     _in(p, dm::DOMAIN) = any(_in(p, e) for e in dm)
 
     @test all(1:100) do _