+++ test/basics.jl
@@ -125,13 +125,6 @@ end
     # Show methods
     @test sprint(summary, italy) == "Italy Borders"
     @test contains(sprint(show, MIME"text/plain"(), italy), ", 1 skipped")
-
-    # centroid
-    dmn = extract_countries("italy; spain")
-    c_ll = centroid(LatLon, dmn)
-    c_cart = centroid(dmn)
-    @test c_cart isa POINT_CART
-    @test centroid(dmn, 1) isa POINT_CART
 end    
 
 @testitem "Coastlines" setup=[setup_basic] begin