+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [0.5.0] - 2025-07-07
+## [0.5.0] - 2025-07-12
 
 This version moved a lot of the internals to the new GeoBasics package.